+++ migrations/20240605192122_stage_instances.sql
@@ -7,7 +7,7 @@ create table if not exists stage_instances
     topic                    varchar(120)         not null,
     privacy_level            int                  not null,
     invite_code              varchar(16)          null,
-    discoverable_disabled    tinyint(1) default 0 not null comment 'deprecated',
+    discoverable_disabled    smallint(1) default 0 not null comment 'deprecated',
     guild_scheduled_event_id varchar(255)         null,
     constraint stage_instances_channels_id_fk
         foreign key (channel_id) references channels (id)
